[ etc ] 무한스크롤 제거 코드
